scripts: retire aur-vercmp-devel
aur-vercmp-devel is just one of ways to determine if a vcs package is
up-to-date. Rather than trying to make aur-vercmp-devel suitable to
everyone, the goal is to find a more general procedure to deal with such
packages.

see #500 #511 and more specifically: issues/511#issuecomment-508938979

For that, let's retire aur-vercmp-devel in favor of documenting
workflows in a new manpage, aurvcs(7).