scripts: use \n after each case) in option parsing
Adding or removing options were often accompanied by formatting changes
to realign everything. The frequent realignments are an issue for blame,
and makes changes harder to review, as purely cosmetic changes show up
mixed with meaningful ones.

This avoids realigning the options whenever an argument is added or
modified.

This commit would be a pure formatting-change-only if it wasn't for the
fact that aur-sync is using extglob to shorten the options. Since this
is no longer a problem, remove that as well.

This commit should be easier to review by looking at the patch after
some pretty-printing. E.g.:

  $ git -c diff.bash.textconv='bash --pretty-print -O extglob' \
        -c core.attributesfile=<(printf 'lib/* diff=bash') \
        show